test : LineSteps 클래스 생성 및 메소드 이동
